requirements.txt
google-generativeai
python-dotenv
requests
beautifulsoup4